design: PostHeader css 분리
